First playable version
This includes:
- A "Map" class that specifies how the order in which rooms are played
- Moving each class originally in the "leeroy.rb" file to it's own file
- Moving the room classes to the "rooms" directory
- Assorted bugfixes
