+++ rooms/chapelRoom.rb
