+++ rooms/ogre.rb
