+++ rooms/silverfish.rb
